// constants/Enums.ts
export enum CategoryColor {
  gray = 1,
  red = 2,
  orange = 3,
  yellow = 4,
  green = 5,
  cyan = 6,
  blue = 7,
  purple = 8,
  pink = 9,
  black = 10,
}

export const CategoryDisplay = {
  gray: "Cinza",
  red: "Vermelho",
  orange: "Laranja",
  yellow: "Amarelo",
  green: "Verde",
  cyan: "Ciano",
  blue: "Azul",
  purple: "Roxo",
  pink: "Rosa",
  black: "Preto",
};

export enum CategoryIcons {
  question = 1,
  drinks = 2,
  meat = 3,
  portions = 4,
  desserts = 5,
}

export const IconDisplay = {
  question: "Interrogação",
  drinks: "Bebida",
  meat: "Carne",
  portions: "Fritas",
  desserts: "Bala",
};

export enum ItemStatus {
  Pendente = 1,
  Preparando = 2,
  Pronto = 3,
  Entregue = 4,
}